refactor: extract TerminalRenderer to renderer module
Move TerminalRenderer and all rendering helpers (ANSI constants,
markdown formatting, code highlighting, inline span parsing) from
src/im/terminal.rs to src/renderer/terminal.rs, aligning module
structure with docs/design/cli/renderer.md.

- src/renderer/terminal.rs: TerminalRenderer + rendering utilities
- src/im/terminal.rs: TerminalAdapter + TerminalPlugin only
- Tests updated to import from new location

Source: user
Type: refactor
